Move alert tracking to the awareness module
